requirements.txt
Flask>=2.0
Werkzeug>=2.0
google-auth>=2.0
google-genai>=2.3.0
